backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
dj-rest-auth==5.0.2
django-allauth==0.61.1
django-rest-knox==5.0.2 